excel to json to api:
i added to front: Component ExcelUploader -pototype of file browser and upload btn, it also shows the excel json and sends it to backend.

ExcelParser in utils - create a promise that handles the excel file to json object array

backend:

post function for /api/upload - prints the json and gives 200 response.

i also added 'cors" support" because of an error

and middelware to parse json bodies